+++ pi_c_pthreadParameters.c
@@ -8,23 +8,23 @@
 #include <pthread.h>
 #include "microsecondTime.h"
 
-long double sum ;
+double sum ;
 pthread_mutex_t sumMutex ;
 
 typedef struct CalculationParameters {
   long id ;
   long sliceSize ;
-  long double delta ;
+  double delta ;
 } CalculationParameters ;
 
 void * partialSum ( void *const arg  ) {
   const long start = 1 + ( (CalculationParameters *const) arg )->id * ( (CalculationParameters *const) arg )->sliceSize ;
   const long end = ( ( (CalculationParameters *const) arg )->id + 1 ) * ( (CalculationParameters *const) arg )->sliceSize ;
-  const long double delta = ( (CalculationParameters *const) arg )->delta ;
-  long double localSum = 0.0 ;
+  const double delta = ( (CalculationParameters *const) arg )->delta ;
+  double localSum = 0.0 ;
   long i ;
   for ( i = start ; i <= end ; ++i ) {
-    const long double x = ( i - 0.5 ) * delta ;
+    const double x = ( i - 0.5 ) * delta ;
     localSum += 1.0 / ( 1.0 + x * x ) ;
   }
   pthread_mutex_lock ( &sumMutex ) ;
@@ -36,7 +36,7 @@ void * partialSum ( void *const arg  ) {
 
 void execute ( const int numberOfThreads ) {
   const long n = 1000000000l ;
-  const long double delta = 1.0 / n ;
+  const double delta = 1.0 / n ;
   const long long startTimeMicros = microsecondTime ( ) ;
   const long sliceSize  = n / numberOfThreads ;
   pthread_mutex_init ( &sumMutex , NULL ) ;
@@ -56,11 +56,11 @@ void execute ( const int numberOfThreads ) {
   pthread_attr_destroy ( &attributes ) ;
   int status ;
   for ( i = 0 ; i < numberOfThreads ; ++i ) { pthread_join ( threads[i] , (void **) &status ) ; }
-  const long double pi = 4.0 * sum * delta ;
-  const long double elapseTime = ( microsecondTime ( ) - startTimeMicros ) / 1e6 ;
-  printf ( "==== C PThread parameters pi = %.25Lf\n" , pi ) ;
+  const double pi = 4.0 * sum * delta ;
+  const double elapseTime = ( microsecondTime ( ) - startTimeMicros ) / 1e6 ;
+  printf ( "==== C PThread parameters pi = %.18lf\n" , pi ) ;
   printf ( "==== C PThread parameters iteration count = %ld\n" ,  n ) ;
-  printf ( "==== C PThread parameters elapse = %Lf\n" , elapseTime ) ;
+  printf ( "==== C PThread parameters elapse = %lf\n" , elapseTime ) ;
   printf ( "==== C PThread parameters thread count = %d\n" , numberOfThreads ) ;
 }
