+++ pi_c_sequential.c
@@ -1,7 +1,7 @@
 /*
  *  A C program to calculate Pi using quadrature as a sequential algorithm.
  *
- *  Copyright © 2008-9 Russel Winder
+ *  Copyright © 2008-10 Russel Winder
  */
 
 #include <stdio.h>
@@ -9,18 +9,18 @@
 
 int main ( ) {
   const long n = 1000000000l ;
-  const long double delta = 1.0 / n ;
+  const double delta = 1.0 / n ;
   const long long startTimeMicros = microsecondTime ( ) ;
-  long double sum = 0.0 ;
+  double sum = 0.0 ;
   long i ;
   for ( i = 1 ; i <= n ; ++i ) {
-    const long double x = ( i - 0.5 ) * delta ;
+    const double x = ( i - 0.5 ) * delta ;
     sum += 1.0 / ( 1.0 + x * x ) ;
   }
-  const long double pi = 4.0 * sum * delta ;
-  const long double elapseTime = ( microsecondTime ( ) - startTimeMicros ) / 1e6 ;
-  printf ( "==== C Sequential pi = %.18Lf\n" , pi ) ;
+  const double pi = 4.0 * sum * delta ;
+  const double elapseTime = ( microsecondTime ( ) - startTimeMicros ) / 1e6 ;
+  printf ( "==== C Sequential pi = %.18lf\n" , pi ) ;
   printf ( "==== C Sequential iteration count = %ld\n" ,  n ) ;
-  printf ( "==== C Sequential elapse = %Lf\n" , elapseTime ) ;
+  printf ( "==== C Sequential elapse = %lf\n" , elapseTime ) ;
   return 0 ;
 }
